+++ src/FrontDoor/FrontDoor/ChangeLog.md
@@ -18,6 +18,8 @@
         - Additional information about change #1
 -->
 ## Upcoming Release
+* Added WAF managed rules exclusion support
+* Add SocketAddr to auto-complete
 
 ## Version 1.2.0
 * Add MinimumTlsVersion parameter to Enable-AzFrontDoorCustomDomainHttps and New-AzFrontDoorFrontendEndpointObject